requirements.txt
appsignal
fastapi[all]
opentelemetry-instrumentation-fastapi
aiohttp
requests
sqlalchemy
psycopg2-binary